refactor(explore): split explore_screen.dart by tab
Breaks the 1,258-line explore_screen.dart into a thin Page/View plus
per-tab and per-widget files under lib/screens/explore/, coordinated by
a new ExploreTabsCubit.

- ExploreTabsCubit + ExploreTabsState own optional-tab availability, the
  ordered tab-name list, and name<->index conversion, so callers never
  reason about raw indices that shift as tabs appear/disappear.
- ExploreScreen becomes a thin Page (75 lines) that provides the cubit;
  ExploreView holds the tab controller lifecycle and scaffold.
- Extracts ExploreListsTab (+ subscribed-lists section), ExploreTabBar,
  ExploreTabView, ExploreNewVideosBanner, and ExploreFeedContent into
  their own files.
- Moves the screen to lib/screens/explore/explore_screen.dart and updates
  all importers; preserves the /explore/tab/:name route contract and the
  ExploreScreen static route helpers.
- Removes dead public State methods (onScreenVisible/onScreenHidden/
  showHashtagVideos/currentHashtag/customTitle/onTabTapped) that had no
  remaining callers.
- Adds ExploreTabsCubit unit tests; existing explore screen/router tests
  pass unchanged except the one assertion updated for the Page/View split.

Every resulting file is comfortably within the size targets (Page 75
lines; View 383; lists tab 429; others < 100).

Closes #4515
