+++ mobile/lib/router/providers/redirect_provider.dart
@@ -5,7 +5,7 @@ import 'dart:convert';
 
 import 'package:openvine/providers/shared_preferences_provider.dart';
 import 'package:openvine/screens/auth/welcome_screen.dart';
-import 'package:openvine/screens/explore_screen.dart';
+import 'package:openvine/screens/explore/explore_screen.dart';
 import 'package:riverpod/src/providers/provider.dart';
 import 'package:unified_logger/unified_logger.dart';
