+++ mobile/test/providers/explore_active_video_test.dart
@@ -10,7 +10,7 @@ import 'package:openvine/providers/active_video_provider.dart';
 import 'package:openvine/providers/app_lifecycle_provider.dart';
 import 'package:openvine/providers/video_events_providers.dart';
 import 'package:openvine/router/router.dart';
-import 'package:openvine/screens/explore_screen.dart';
+import 'package:openvine/screens/explore/explore_screen.dart';
 
 void main() {
   group('Explore Active Video Provider', () {
