+++ mobile/test/providers/video_playback_stop_integration_test.dart
@@ -10,7 +10,7 @@ import 'package:openvine/providers/active_video_provider.dart';
 import 'package:openvine/providers/app_foreground_provider.dart';
 import 'package:openvine/providers/route_feed_providers.dart';
 import 'package:openvine/router/router.dart';
-import 'package:openvine/screens/explore_screen.dart';
+import 'package:openvine/screens/explore/explore_screen.dart';
 import 'package:openvine/state/video_feed_state.dart';
 
 void main() {
