+++ mobile/test/screens/explore_screen_apps_tab_test.dart
@@ -12,7 +12,7 @@ import 'package:openvine/providers/for_you_provider.dart';
 import 'package:openvine/providers/list_providers.dart';
 import 'package:openvine/providers/route_feed_providers.dart';
 import 'package:openvine/router/router.dart';
-import 'package:openvine/screens/explore_screen.dart';
+import 'package:openvine/screens/explore/explore_screen.dart';
 import 'package:openvine/screens/user_list_people_screen.dart';
 import 'package:openvine/services/curated_list_service.dart';
 import 'package:openvine/services/video_event_service.dart';
