+++ mobile/test/screens/explore_screen_video_display_test.dart
@@ -14,7 +14,7 @@ import 'package:openvine/providers/nostr_client_provider.dart';
 import 'package:openvine/providers/seen_videos_notifier.dart';
 import 'package:openvine/providers/video_events_providers.dart';
 import 'package:openvine/router/router.dart';
-import 'package:openvine/screens/explore_screen.dart';
+import 'package:openvine/screens/explore/explore_screen.dart';
 import 'package:openvine/services/video_event_service.dart';
 
 import '../test_data/video_test_data.dart';
